+++ src/main/kotlin/com/smody/book/auth/dto/LoginResponse.kt
@@ -1,6 +1,6 @@
-package com.smody.book.security.dto
+package com.smody.book.auth.dto
 
-import com.smody.book.security.OAuthPrincipal
+import com.smody.book.auth.security.OAuthPrincipal
 
 data class LoginResponse(val memberId: Long, val accessToken: String) {
     constructor(oAuthPrincipal: OAuthPrincipal) : this(oAuthPrincipal.memberId, oAuthPrincipal.accessToken)
