+++ src/main/kotlin/com/smody/book/auth/security/OAuthPrincipal.kt
@@ -1,4 +1,4 @@
-package com.smody.book.security
+package com.smody.book.auth.security
 
 import org.springframework.security.core.GrantedAuthority
 import org.springframework.security.oauth2.core.user.OAuth2User
