+++ src/test/kotlin/com/smody/book/security/authentication/OAuthLoginSuccessHandlerTest.kt
@@ -1,7 +1,8 @@
 package com.smody.book.security.authentication
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.smody.book.security.OAuthPrincipal
+import com.smody.book.auth.security.OAuthPrincipal
+import com.smody.book.auth.security.authentication.OAuthLoginSuccessHandler
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.junit.jupiter.api.DisplayName
